refactor: add controller groups in routes
